README: Translate README to use Asciidoctor instead of markdown
Rename file to README.adoc in order to allow Github to automatically render the
file when displaying it at repository's page.

Translate markdown to asciidoctor. It's not a big change, but it includes some
formatting like source code highlighting.

Remove some future tense used in the text, and fixed some other sentences.

Signed-off-by: Daiane Angolini <daiane.angolini@freescale.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>